Create a standardised leaflet base map for two layers side-by-side

Create a leaflet map upon which to add layers through leaflet::add functions or leafletProxy. Uses leaflet.extras2::sidebyside to add a vertical slider to view two layers side-by-side. Add additional layers to a single side by specifying options = pathOptions(pane = 'left/right'), or add layers above the slider by using the panes parameters with zIndex >1.

build_basemap_sidebyside(
  panes = NULL,
  panes_zIndex = NULL,
  basemap = c("OpenStreetMap", "ImageMapSA", "Esri imagery"),
  ...
)

Arguments

panes

Character vector of names to add as additional height-ordered layers. Call in added layers with options=pathOptions(pane = 'blah')

panes_zIndex

Numeric vector of panes levels. Must be equal length to panes.

basemap

Name of basemap layer to use. Note that sidebyside doesn't support layerControl for basemaps, so only one value is used. Supports default OpenStreetMap & ESRI World Imagery, and ImageMapSA via its public URL (which reverts to ESRI imagery outside its extent).

...

Passed to initial leaflet(), eg. for setting height.

Value

A leaflet map, as an HTML widget object.

Details

Includes fewer options than build_basemap, which is better for adding multiple layers.