Make a zipfile of shapefile parts to be downloaded.

Create the components of one or more shapefiles, and zip them to a single file that is ready for shiny::downloadHandler. Assign to an object (e.g. `zipfile`) within the function in the content argument of downloadHandler, then use file.copy(zipfile, file) (see example).

make_shp_download(
  data,
  layer_names,
  zip_name = if (!multiple) layer_names,
  multiple = FALSE
)

Arguments

data

Data to be downloaded, usually a reactive. If multiple = TRUE, must be a list.

layer_names

Name of the shapefile/s, ie will output layer_name.shp, layer_name.dbf, etc. If multiple = TRUE, should be same length as data list - one name for each shapefile.

zip_name

Name of the .zip folder downloaded. MUST match the name provided to the filename argument of shiny::download_handler.

multiple

Download contains multiple shapefiles? Check requirements for data and layer_names if TRUE.

Value

A temp path to the .zip file ready to be downloaded.

Details

Replaces envShiny::download_shp (see [envShiny #2](https://github.com/dew-landscapes/envShiny/issues/2)).

Examples

if (FALSE) { # \dontrun{
output$download <- downloadHandler(
filename = function() {
  paste0("download_", input$taxa, ".zip")
},
content = function(file) {
  zipfile <- make_shp_download(data = data(),
                               layer_names = paste0("data_", input$taxa),
                               zip_name = paste0("download_", input$taxa),
                               multiple = FALSE)
  file.copy(zipfile, file)
}
)
} # }